codial-discord/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codial-discord"
version = "0.1.0"
description = "Codial Discord gateway bot"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "httpx>=0.27.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.3.0",
  "pynacl>=1.5.0",
  "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
  "pytest-cov>=5.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0",
]

[project.scripts]
codial-discord = "codial_discord.cli:main"
codial-discord-dev = "codial_discord.cli:main_dev"

[tool.setuptools.packages.find]
where = ["."]
include = ["codial_discord*"]
